requirements.txt
flet~=0.28.3
networkx~=3.5
mysql-connector-python~=8.4.0
